refactor: simplify module export
