+++ src/infrastructure/http-clients/maps/map-box/mapbox-http-client-adapter.spec.ts
@@ -1,7 +1,7 @@
 import { Coordinate } from '@domain/models'
 import { faker } from '@faker-js/faker'
 import { httpMapClientsConstants } from '@infrastructure/http-clients/settings'
-import { mockForwardGeoCodingResponse } from '@infrastructure/http-clients/test/mock-geocoding'
+import { mockForwardGeoCodingResponse } from '@infrastructure/http-clients/test'
 import { MapBoxHttpClientAdapter } from './mapbox-http-client-adapter'
 
 const unmockedFetch = global.fetch
@@ -38,6 +38,21 @@ describe('MapBoxHttpClientAdapter', () => {
 
     await sut.retrieveCoordinates(address)
     expect(fetchSpy).toHaveBeenCalledWith(input, options)
+  })
+
+  test('Should return latitude and longitude if feature list is not empty for given address', async () => {
+    const sut = new MapBoxHttpClientAdapter()
+
+    const address = faker.address.city()
+
+    const response = mockForwardGeoCodingResponse()
+
+    const coordinates = await sut.retrieveCoordinates(address)
+
+    expect(coordinates).toEqual<Coordinate>({
+      latitude: response.features[0].center[0],
+      longitude: response.features[0].center[1]
+    })
   })
 
   test('Should return latitude and longitude if feature list is not empty for given address', async () => {
