+++ src/infrastructure/http-clients/test/index.ts
@@ -1,0 +1,1 @@
+export * from './mock-geocoding'
